backend/requirements.txt
asgiref==3.10.0
boto3==1.40.71
botocore==1.40.71
coverage==7.11.3
Django==5.2.8
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
gunicorn==23.0.0
iniconfig==2.3.0
jmespath==1.0.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
